updating get_y_class and testing codeclimate
